apps/ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
torch==2.4.0
torchvision==0.19.0
pillow==10.4.0
numpy==1.26.4
python-multipart==0.0.9
pydantic==2.8.0
httpx==0.27.0
python-dotenv==1.0.1
